# Fix unicode surrogate pairs not being handled correctly