--- a/src/main/php/text/json/StreamOutput.class.php
+++ b/src/main/php/text/json/StreamOutput.class.php
@@ -6,7 +6,7 @@ use io\streams\OutputStream;
  * Writes JSON to a given output stream
  *
  * ```php
- * $json= new StreamOutput((new stream('output.json'))->getOutputStream()));
+ * $json= new StreamOutput((new File('output.json'))->out()));
  * $json->write('Hello World');
  * ```
  *