## Changes

Fixed encoding and decoding of the following:

| String | JSON | Comment |
| -------- | --------- | ------------- |
| Poop = 💩 | `"Poop = \ud83d\udca9"` | *Code points are encoded as 2 units called surrogate pairs*

See xp-framework/webservices#1

## Performance

Performance is hit in the sub-millisecond area when comparing `$hex > 0xd800 && $hex < 0xdfff`.

Dimension | Before | After |
| ----------- | ----------- | --------- |
| Runtime | 1.473 seconds | 1.478 seconds |
| Iteration | 14.7 ms per iteration, result = 30 | 14.8 ms per iteration, result = 30 |
| Memory | 874.086 kB memory used | 875.523 kB memory used |
| Peak memory | 920.328 kB peak | 921.766 kB peak |

*This is parsing a 158791 bytes JSON file 100 times.*